fix responsive false on box
