+++ package.json
@@ -78,7 +78,7 @@
     "@storybook/react": "^5.0.1",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
-    "babel-jest": "^23.0.1",
+    "babel-jest": "^24.8.0",
     "babel-loader": "^8.0.0-beta.6",
     "babel-plugin-styled-components": "^1.6.4",
     "babel-plugin-transform-imports": "^1.5.0",
@@ -102,8 +102,8 @@
     "grommet-theme-hp": "^0.1.1",
     "grommet-theme-hpe": "^0.4.0",
     "grommet-theme-v1": "^0.1.1",
-    "jest": "^23.1.0",
-    "jest-cli": "^23.1.0",
+    "jest": "^24.8.0",
+    "jest-cli": "^24.8.0",
     "jest-styled-components": "^6.2.0",
     "jsdom": "^13.1.0",
     "pre-commit": "^1.2.2",
