+++ src/js/components/Box/stories/Background.js
@@ -1,0 +1,43 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Grommet, Box } from 'grommet';
+import { grommet } from '../../../themes';
+
+const BackgroundBox = () => (
+  <Grommet theme={grommet}>
+    <Box pad="small" gap="small" align="start">
+      <Box
+        pad="small"
+        background={{ color: 'brand', opacity: true }}
+        elevation="large"
+      >
+        brand opacity
+      </Box>
+      <Box pad="small" background="brand" elevation="large">
+        brand
+      </Box>
+      <Box
+        pad="small"
+        background={{
+          image:
+            'url(http://librelogo.org/wp-content/uploads/2014/04/gradient2.png)',
+        }}
+      >
+        image
+      </Box>
+      <Box
+        pad="small"
+        background={{
+          color: 'accent-2',
+          image:
+            'url(http://librelogo.org/wp-content/uploads/2014/04/gradient2.png)',
+        }}
+      >
+        image + color
+      </Box>
+    </Box>
+  </Grommet>
+);
+
+storiesOf('Box', module).add('Background', () => <BackgroundBox />);
